New feature: extrude() builtin
This new builtin module takes a collection of 2D children and applies
an outer "skin" to them in the same manner as the other extrude ops.
The other ops can now be thought of as optimizations that run an
extrude() on specifically constructed list of children.

The extrude() feature is initially very restrictive; it requires that
each child element must be entirely in the +Z space of the previous
element.  This restriction could be relaxed later with smarter sanity
checking and more extrusion logic.